move statistics out into separate crate 'flow' remove sendable trait from sentinel introduce pure sentinel for identical messages
